third_party/WebKit/LayoutTests/webaudio/realtimeanalyser-byte-data-expected.txt

Test AnalyserNode getByteTimeDomainData

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Index of first sample greater than +1 is greater than or equal to 0.
PASS Index of first sample less than -1 is greater than or equal to 0.
PASS Clip  1.01149: byteData[43] is equal to 255.
PASS Clip -1.42637: byteData[56] is equal to 0.
PASS Byte data is identical to the array [128,130,135,136,140,143,145,149,152,153,159,160,163,167,168,173,...].
PASS successfullyParsed is true

TEST COMPLETE
